service-b/requirements/production.txt
flask==2.0.1
flask-cors==3.0.10
flask-mongoengine==1.0.0
flask-restful==0.3.9
flask-migrate==3.1.0
flask-seeder==1.2.0
faker==8.11.0
requests==2.26.0
python-dotenv==0.19.0
